+++ src/app/inobeta-ui/ui/table/table.component.ts
@@ -88,6 +88,38 @@ import { formatDate } from '@angular/common';
               (delete)="delete.emit($event)"
             >
             </tr>
+            <tr
+              class="table-row ib-table-row-totals-label"
+              *ngIf="hasTotals()"
+              >
+                <td
+                  [attr.colspan]="titles.length + templateButtons.length + (selectableRows ? 1 : 0) + (hasEdit ? 1 : 0) + (hasDelete ? 1 : 0)"
+                >
+                  {{ 'shared.ibTable.totals' | translate }}
+                </td>
+            </tr>
+            <tr
+              class="table-row ib-table-row-totals"
+              *ngIf="hasTotals()"
+              >
+              <td *ngIf="selectableRows"></td>
+              <td
+                *ngFor="let t of titles"
+                style="padding: 10px 15px;"
+                [ngStyle]="{
+                   'text-align': 'right'
+                }"
+                class="ib-table-column-type-number"
+              >
+              <span *ngIf="t.showTotalSum" style="white-space:nowrap;">
+                {{ 'shared.ibTable.totalPerPage' | translate }} {{ getTotalsOfPage(t.key) | number:t.format:'it'}}<br />
+                {{ 'shared.ibTable.totalAllPages' | translate }} {{ getTotalsAll(t.key) | number:t.format:'it'}}
+              </span>
+              </td>
+              <td *ngFor="let i of templateButtons"></td>
+              <td *ngIf="hasEdit"></td>
+              <td *ngIf="hasDelete"></td>
+            </tr>
           </tbody>
 
           <tr *ngIf="sortedData.length === 0">
@@ -184,6 +216,7 @@ export class IbTableComponent implements OnChanges {
   typeEnum = IbTableTitlesTypes;
   alignEnum = IbTableCellAligns;
   sortedData: IbTableItem[];
+  filteredData: IbTableItem[];
   currentPagination: any = {};
   columnFilter = {};
   numOfElements = 0;
@@ -434,6 +467,7 @@ export class IbTableComponent implements OnChanges {
       }
       paginatedData.push(this.sortedData[i]);
     }
+    this.filteredData = this.sortedData.slice();
     this.sortedData = paginatedData;
   }
 
@@ -542,6 +576,26 @@ export class IbTableComponent implements OnChanges {
         document.body.removeChild(link);
       }
     }
+  }
+
+  getTotalsOfPage(key){
+    return this.sortedData.reduce((acc, el) => {
+      acc += el[key];
+      return acc;
+    }, 0);
+  }
+
+  getTotalsAll(key){
+    return this.filteredData.reduce((acc, el) => {
+      acc += el[key];
+      return acc;
+    }, 0);
+  }
+  hasTotals(){
+    for (const t of this.titles){
+      if (t.showTotalSum) { return true; }
+    }
+    return false;
   }
 }