+++ src/app/inobeta-ui/ui/table/table.component.ts
@@ -39,7 +39,7 @@ import { formatDate } from '@angular/common';
         (export)="export($event)"
       >
       </div>
-      <div>
+      <div class="ib-table-container">
         <table
           matSort
           (matSortChange)="sortData($event)"