+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - DEVK-108 Fixed issue on date filter when using datetime data
 - DEVK-84 Removed deprecated errors field from IbFormControlBase
 - Fixed pipeline issue detected on job <https://gitlab.com/Inobeta/Interni/DevKit/inobeta-ui/-/jobs/1351671095>
+- DEVK-127 showTotalSum option added to IbTableTitles
 
 ## [12.0.0]