+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - DEVK-84 Removed deprecated errors field from IbFormControlBase
 - Fixed pipeline issue detected on job <https://gitlab.com/Inobeta/Interni/DevKit/inobeta-ui/-/jobs/1351671095>
 - DEVK-127 showTotalSum option added to IbTableTitles
+- DEVK-128 fixed bug on change trigger for IbMatDropdownControl when multiple: true
 
 ## [12.0.0]